rfp_ai_system/requirements.txt
langgraph
langchain
google-generativeai
beautifulsoup4
requests
pandas
openpyxl
python-dotenv
tabulate
